refactor: parameter 비어있을 경우 에러처리

+++ backend/src/user/user.service.ts
@@ -8,27 +8,27 @@ import {
   NotFoundException,
   UnauthorizedException,
 } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { JwtService } from '@nestjs/jwt';
+import {InjectRepository} from '@nestjs/typeorm';
+import {JwtService} from '@nestjs/jwt';
 
 import axios from 'axios';
 import * as bcrypt from 'bcryptjs';
 
 import * as fs from 'fs';
 
-import { User } from './user.entitiy';
-import { UserDto } from './dto/user.dto';
-import { UserAuthDto } from './dto/userAuth.dto';
-import { UserRepository } from './user.repository';
-import { IsUserAuthRepository } from 'src/signup/signup.repository';
-import { Payload } from './payload';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { SocketArray } from '@/global-variable/global.socket';
-import { SignUpService } from '@/signup/signup.service';
-import { OAuthUser } from '@/types/user/oauth';
-import { Friend } from '@/global-variable/global.friend';
-import { DataSource, QueryRunner } from 'typeorm';
-import { Block } from '@/global-variable/global.block';
+import {User} from './user.entitiy';
+import {UserDto} from './dto/user.dto';
+import {UserAuthDto} from './dto/userAuth.dto';
+import {UserRepository} from './user.repository';
+import {IsUserAuthRepository} from 'src/signup/signup.repository';
+import {Payload} from './payload';
+import {UpdateUserDto} from './dto/update-user.dto';
+import {SocketArray} from '@/global-variable/global.socket';
+import {SignUpService} from '@/signup/signup.service';
+import {OAuthUser} from '@/types/user/oauth';
+import {Friend} from '@/global-variable/global.friend';
+import {DataSource, QueryRunner} from 'typeorm';
+import {Block} from '@/global-variable/global.block';
 
 @Injectable()
 export class UserService {
@@ -40,13 +40,13 @@ export class UserService {
     private friend: Friend,
     private block: Block,
     private dataSource: DataSource
-  ) { }
+  ) {}
 
   async findUser(user_id: string): Promise<User> {
     if (!user_id) {
-      throw new BadRequestException('유저아이디를 입력해주세요.');
+      throw new BadRequestException('아이디를 입력해주세요.');
     }
-    const found = await this.userRepository.findOneBy({ user_id });
+    const found = await this.userRepository.findOneBy({user_id});
     if (!found) {
       throw new NotFoundException(`${user_id}는 유저가 아닙니다.`);
     }
@@ -54,6 +54,9 @@ export class UserService {
   }
 
   async remove(user_id: string): Promise<void> {
+    if (!user_id) {
+      throw new BadRequestException('아이디를 입력해주세요.');
+    }
     const result = await this.userRepository.delete(user_id);
     if (result.affected === 0) {
       throw new NotFoundException(`${user_id}는 유저가 아닙니다.`);
@@ -72,13 +75,14 @@ export class UserService {
   }
 
   async changePW(user_id: string, userDto: UpdateUserDto): Promise<void> {
+    if (!user_id) {
+      throw new BadRequestException('아이디를 입력해주세요.');
+    }
     const user = await this.findUser(user_id);
     if (user) {
-      //TODO: 주석 제거하기
       const salt = await bcrypt.genSalt();
       const hashedPassword = await bcrypt.hash(userDto.user_pw, salt);
-      // user.user_pw = hashedPassword;
-      user.user_pw = userDto.user_pw;
+      user.user_pw = hashedPassword;
       await this.userRepository.save(user);
     } else {
       throw new NotFoundException(`${user_id}는 유저가 아닙니다.`);
@@ -90,19 +94,20 @@ export class UserService {
       throw new ConflictException('로그인중인 유저입니다.');
     }
     const user = await this.findUser(userAuthDto.user_id);
-    if (user && userAuthDto.user_pw === user.user_pw) {
-      // TODO: 주석 제거해서 bcrpyt 적용하기
-      // if (user && (await bcrypt.compare(userAuthDto.user_pw, user.user_pw))) {
+    if (!user) {
+      throw new NotFoundException('유저가 아닙니다.');
+    }
+    if (await bcrypt.compare(userAuthDto.user_pw, user.user_pw)) {
       // user token create. (secret + Payload)
       if (user.is_2fa_enabled === false) {
         if (user.two_factor_auth_secret !== null) {
           // 2FA 미사용 유저의 시크릿 키 삭제
           await this.userRepository.update(
-            { user_id: userAuthDto.user_id },
-            { two_factor_auth_secret: null }
+            {user_id: userAuthDto.user_id},
+            {two_factor_auth_secret: null}
           );
         }
-        const payload: Payload = { user_id: userAuthDto.user_id };
+        const payload: Payload = {user_id: userAuthDto.user_id};
         const accessToken = this.generateAccessToken(payload);
         return accessToken;
       } else {
@@ -110,15 +115,18 @@ export class UserService {
         return HttpStatus.ACCEPTED;
       }
     }
-    throw new NotFoundException('유저가 아닙니다.');
+    throw new UnauthorizedException('비밀번호가 틀립니다.');
   }
 
   async getOAuthUser(code: string): Promise<string | OAuthUser> {
+    if (!code) {
+      throw new BadRequestException('code가 비어있습니다.');
+    }
     const api_uri = process.env.INTRA_API_URI;
     const accessToken = await this.signupService.getAccessToken(code);
 
     const response = await axios.get(api_uri, {
-      headers: { Authorization: `Bearer ${accessToken}` },
+      headers: {Authorization: `Bearer ${accessToken}`},
     });
     const user = await this.userRepository.findOneBy({
       user_id: response.data.login,
@@ -126,7 +134,7 @@ export class UserService {
     if (user) {
       if (user.is_2fa_enabled === false) {
         try {
-          const payload: Payload = { user_id: response.data.login };
+          const payload: Payload = {user_id: response.data.login};
           const accessToken = this.generateAccessToken(payload);
           return accessToken;
         } catch (error) {
@@ -134,7 +142,7 @@ export class UserService {
           throw new InternalServerErrorException('서버에러가 발생했습니다.');
         }
       } else {
-        return { status: HttpStatus.ACCEPTED, user_id: user.user_id };
+        return {status: HttpStatus.ACCEPTED, user_id: user.user_id};
       }
     } else {
       throw new NotFoundException('회원가입이 필요합니다.');
@@ -146,6 +154,9 @@ export class UserService {
     file: Express.Multer.File,
     userID: string
   ): Promise<void> {
+    if (!userID) {
+      throw new BadRequestException('아이디를 입력해주세요.');
+    }
     const user = await this.findUser(userID);
     if (!user) {
       throw new NotFoundException('유저가 아닙니다.');
@@ -155,30 +166,16 @@ export class UserService {
     await query_runner.startTransaction();
     try {
       if (userDto.user_nickname) {
-        // await this.userRepository.update(
-        //   {user_id: userID},
-        //   {
-        //     user_nickname: userDto.user_nickname,
-        //   }
-        // );
         await query_runner.manager.getRepository(User).update(
-          { user_id: userID },
+          {user_id: userID},
           {
             user_nickname: userDto.user_nickname,
           }
         );
       }
       if (file) {
-        // await this.userRepository.update(
-        //   {user_id: userID},
-        //   {
-        //     user_image:
-        //       `${process.env.NEXT_PUBLIC_BACKEND_SERVER}` +
-        //       file.path.substr(11),
-        //   }
-        // );
         await query_runner.manager.getRepository(User).update(
-          { user_id: userID },
+          {user_id: userID},
           {
             user_image:
               `${process.env.NEXT_PUBLIC_BACKEND_SERVER}` +
@@ -221,7 +218,7 @@ export class UserService {
 
   async setTwoFactorAuthSecret(userID: string, secret: string) {
     return this.userRepository.update(
-      { user_id: userID },
+      {user_id: userID},
       {
         two_factor_auth_secret: secret,
       }
@@ -230,7 +227,7 @@ export class UserService {
 
   async turnOnTwoFactorAuth(userID: string) {
     return await this.userRepository.update(
-      { user_id: userID },
+      {user_id: userID},
       {
         is_2fa_enabled: true,
       }
@@ -239,7 +236,7 @@ export class UserService {
 
   async turnOffTwoFactorAuth(userID: string) {
     return await this.userRepository.update(
-      { user_id: userID },
+      {user_id: userID},
       {
         two_factor_auth_secret: null,
         is_2fa_enabled: false,
@@ -248,7 +245,7 @@ export class UserService {
   }
 
   async getUser(userID: string): Promise<Partial<User>> {
-    const user = await this.userRepository.findOneBy({ user_id: userID });
+    const user = await this.userRepository.findOneBy({user_id: userID});
     const userData: Partial<User> = {
       user_id: user.user_id,
       user_nickname: user.user_nickname,
